Remove UVM/container cloning functionality
This change completely removes the support for cloning/late cloning that
was added a few years ago. The reasoning behind this is as follows:
- There are no plans to utilize cloning at this point in time.
- The cloning support required extensive/invasive changes across many
  parts of the shim. This has made future changes and refactorings more
  difficult in some cases. While these changes are still possible, it
  seems like an unnecessary burden if we are not going to use cloning.
- The cloning functionality was never actually utilized, and thus may
  still have had fixes needed to be production-ready.

If cloning is needed again in the future, we will be able to revert this
commit to add it back.
