+++ cmd/containerd-shim-runhcs-v1/task_wcow_podsandbox.go
@@ -9,7 +9,6 @@ import (
 
 	"github.com/Microsoft/hcsshim/cmd/containerd-shim-runhcs-v1/options"
 	"github.com/Microsoft/hcsshim/cmd/containerd-shim-runhcs-v1/stats"
-	"github.com/Microsoft/hcsshim/internal/clone"
 	"github.com/Microsoft/hcsshim/internal/cmd"
 	"github.com/Microsoft/hcsshim/internal/log"
 	"github.com/Microsoft/hcsshim/internal/oc"
@@ -191,10 +190,6 @@ func (wpst *wcowPodSandboxTask) close(ctx context.Context) {
 
 			if err := wpst.host.Close(); err != nil {
 				log.G(ctx).WithError(err).Error("failed host vm shutdown")
-			}
-			// cleanup template state if any exists
-			if err := clone.RemoveSavedTemplateConfig(wpst.host.ID()); err != nil {
-				log.G(ctx).WithError(err).Error("failed to cleanup template config state for vm")
 			}
 		}
 		// Send the `init` exec exit notification always.
