--- internal/clone/doc.go
@@ -1,1 +1,0 @@
-package clone
