+++ internal/uvm/create_wcow.go
@@ -34,22 +34,6 @@ type OptionsWCOW struct {
 
 	LayerFolders []string // Set of folders for base layers and scratch. Ordered from top most read-only through base read-only layer, followed by scratch
 
-	// IsTemplate specifies if this UVM will be saved as a template in future. Setting
-	// this option will also enable some VSMB Options during UVM creation that allow
-	// template creation.
-	IsTemplate bool
-
-	// IsClone specifies if this UVM should be created by cloning a template. If
-	// IsClone is true then a valid UVMTemplateConfig struct must be passed in the
-	// `TemplateConfig` field.
-	IsClone bool
-
-	// TemplateConfig is only used during clone creation. If a uvm is
-	// being cloned then this TemplateConfig struct must be passed
-	// which holds all the information about the template from
-	// which this clone should be created.
-	TemplateConfig *UVMTemplateConfig
-
 	// NoDirectMap specifies that no direct mapping should be used for any VSMBs added to the UVM
 	NoDirectMap bool
 
@@ -244,8 +228,6 @@ func prepareConfigDoc(ctx context.Context, uvm *UtilityVM, opts *OptionsWCOW, uv
 }
 
 // CreateWCOW creates an HCS compute system representing a utility VM.
-// The HCS Compute system can either be created from scratch or can be cloned from a
-// template.
 //
 // WCOW Notes:
 //   - The scratch is always attached to SCSI 0:0
@@ -316,76 +298,42 @@ func CreateWCOW(ctx context.Context, opts *OptionsWCOW) (_ *UtilityVM, err error
 		return nil, fmt.Errorf("error in preparing config doc: %s", err)
 	}
 
-	if !opts.IsClone {
-		// Create sandbox.vhdx in the scratch folder based on the template, granting the correct permissions to it
-		scratchPath := filepath.Join(scratchFolder, "sandbox.vhdx")
-		if _, err := os.Stat(scratchPath); os.IsNotExist(err) {
-			if err := wcow.CreateUVMScratch(ctx, uvmFolder, scratchFolder, uvm.id); err != nil {
-				return nil, fmt.Errorf("failed to create scratch: %s", err)
-			}
-		} else {
-			// Sandbox.vhdx exists, just need to grant vm access to it.
-			if err := wclayer.GrantVmAccess(ctx, uvm.id, scratchPath); err != nil {
-				return nil, errors.Wrap(err, "failed to grant vm access to scratch")
-			}
+	// Create sandbox.vhdx in the scratch folder based on the template, granting the correct permissions to it
+	scratchPath := filepath.Join(scratchFolder, "sandbox.vhdx")
+	if _, err := os.Stat(scratchPath); os.IsNotExist(err) {
+		if err := wcow.CreateUVMScratch(ctx, uvmFolder, scratchFolder, uvm.id); err != nil {
+			return nil, fmt.Errorf("failed to create scratch: %s", err)
 		}
-
-		doc.VirtualMachine.Devices.Scsi = map[string]hcsschema.Scsi{}
-		for i := 0; i < int(uvm.scsiControllerCount); i++ {
-			doc.VirtualMachine.Devices.Scsi[guestrequest.ScsiControllerGuids[i]] = hcsschema.Scsi{
-				Attachments: make(map[string]hcsschema.Attachment),
-			}
-		}
-
-		doc.VirtualMachine.Devices.Scsi[guestrequest.ScsiControllerGuids[0]].Attachments["0"] = hcsschema.Attachment{
-
-			Path:  scratchPath,
-			Type_: "VirtualDisk",
-		}
-
-		uvm.scsiLocations[0][0] = newSCSIMount(uvm,
-			doc.VirtualMachine.Devices.Scsi[guestrequest.ScsiControllerGuids[0]].Attachments["0"].Path,
-			"",
-			doc.VirtualMachine.Devices.Scsi[guestrequest.ScsiControllerGuids[0]].Attachments["0"].Type_,
-			"",
-			1,
-			0,
-			0,
-			false,
-			false)
 	} else {
-		doc.VirtualMachine.RestoreState = &hcsschema.RestoreState{}
-		doc.VirtualMachine.RestoreState.TemplateSystemId = opts.TemplateConfig.UVMID
-
-		for _, cloneableResource := range opts.TemplateConfig.Resources {
-			err = cloneableResource.Clone(ctx, uvm, &cloneData{
-				doc:           doc,
-				scratchFolder: scratchFolder,
-				uvmID:         opts.ID,
-			})
-			if err != nil {
-				return nil, fmt.Errorf("failed while cloning: %s", err)
-			}
+		// Sandbox.vhdx exists, just need to grant vm access to it.
+		if err := wclayer.GrantVmAccess(ctx, uvm.id, scratchPath); err != nil {
+			return nil, errors.Wrap(err, "failed to grant vm access to scratch")
 		}
-
-		// we add default clone namespace for each clone. Include it here.
-		if uvm.namespaces == nil {
-			uvm.namespaces = make(map[string]*namespaceInfo)
-		}
-		uvm.namespaces[DefaultCloneNetworkNamespaceID] = &namespaceInfo{
-			nics: make(map[string]*nicInfo),
-		}
-		uvm.IsClone = true
-		uvm.TemplateID = opts.TemplateConfig.UVMID
 	}
 
-	// Add appropriate VSMB share options if this UVM needs to be saved as a template
-	if opts.IsTemplate {
-		for _, share := range doc.VirtualMachine.Devices.VirtualSmb.Shares {
-			uvm.SetSaveableVSMBOptions(share.Options, share.Options.ReadOnly)
+	doc.VirtualMachine.Devices.Scsi = map[string]hcsschema.Scsi{}
+	for i := 0; i < int(uvm.scsiControllerCount); i++ {
+		doc.VirtualMachine.Devices.Scsi[guestrequest.ScsiControllerGuids[i]] = hcsschema.Scsi{
+			Attachments: make(map[string]hcsschema.Attachment),
 		}
-		uvm.IsTemplate = true
 	}
+
+	doc.VirtualMachine.Devices.Scsi[guestrequest.ScsiControllerGuids[0]].Attachments["0"] = hcsschema.Attachment{
+
+		Path:  scratchPath,
+		Type_: "VirtualDisk",
+	}
+
+	uvm.scsiLocations[0][0] = newSCSIMount(uvm,
+		doc.VirtualMachine.Devices.Scsi[guestrequest.ScsiControllerGuids[0]].Attachments["0"].Path,
+		"",
+		doc.VirtualMachine.Devices.Scsi[guestrequest.ScsiControllerGuids[0]].Attachments["0"].Type_,
+		"",
+		1,
+		0,
+		0,
+		false,
+		false)
 
 	err = uvm.create(ctx, doc)
 	if err != nil {
